[MOD] 調整 金額單位為整數

+++ app/src/main/java/com/example/keepaccount/AddItemFragment.kt
@@ -96,7 +96,7 @@ class AddItemFragment : Fragment() {
             viewModel.addNewItem(
                 Item(
                     itemName = binding.itemName.text.toString(),
-                    itemPrice = binding.itemPrice.text.toString().toDouble(),
+                    itemPrice = binding.itemPrice.text.toString().toInt(),
                     itemColorcode = binding.itemColorcode.text.toString(),
                     itemYear = year,
                     itemMonth = month,
