+++ app/src/main/java/com/example/keepaccount/ViewModels/StripViewModel.kt
@@ -74,7 +74,7 @@ class StripViewModel(
             _boardUI.value =
                 StripUIState(
                     budget = budget,
-                    cost = cost.toInt(),
+                    cost = cost,
                 )
         }
     }
